+++ jodd-bean/src/main/java/jodd/introspector/MethodDescriptor.java
@@ -37,13 +37,14 @@ import java.lang.reflect.Type;
  */
 public class MethodDescriptor extends Descriptor implements Getter, Setter {
 
+	private static final MethodParamDescriptor[] NO_PARAMS = new MethodParamDescriptor[0];
+
 	protected final Method method;
 	protected final Type returnType;
 	protected final Class rawReturnType;
 	protected final Class rawReturnComponentType;
 	protected final Class rawReturnKeyComponentType;
-	protected final Class[] rawParameterTypes;
-	protected final Class[] rawParameterComponentTypes;
+	protected final MethodParamDescriptor[] parameters;
 
 	public MethodDescriptor(final ClassDescriptor classDescriptor, final Method method) {
 		super(classDescriptor, ClassUtil.isPublic(method));
@@ -62,17 +63,25 @@ public class MethodDescriptor extends Descriptor implements Getter, Setter {
 
 		ClassUtil.forceAccess(method);
 
-		Type[] params = method.getGenericParameterTypes();
-		Type[] genericParams = method.getGenericParameterTypes();
+		if (method.getParameterCount() == 0) {
+			parameters = NO_PARAMS;
+		}
+		else {
+			parameters = new MethodParamDescriptor[method.getParameterCount()];
 
-		rawParameterTypes = new Class[params.length];
-		rawParameterComponentTypes = genericParams.length == 0 ? null : new Class[params.length];
+			Class[] params = method.getParameterTypes();
+			Type[] genericParams = method.getGenericParameterTypes();
 
-		for (int i = 0; i < params.length; i++) {
-			Type type = params[i];
-			rawParameterTypes[i] = ClassUtil.getRawType(type, classDescriptor.getType());
-			if (rawParameterComponentTypes != null) {
-				rawParameterComponentTypes[i] = ClassUtil.getComponentType(genericParams[i], classDescriptor.getType(), -1);
+			for (int i = 0; i < params.length; i++) {
+				final Class parameterType = params[i];
+				final Class rawParameterType = genericParams.length == 0 ?
+					parameterType :
+					ClassUtil.getRawType(genericParams[i], classDescriptor.getType());
+				final Class rawParameterComponentType = genericParams.length == 0 ?
+					null :
+					ClassUtil.getComponentType(genericParams[i], classDescriptor.getType(), -1);
+
+				parameters[i] = new MethodParamDescriptor(parameterType, rawParameterType, rawParameterComponentType);
 			}
 		}
 	}
@@ -126,18 +135,17 @@ public class MethodDescriptor extends Descriptor implements Getter, Setter {
 	}
 
 	/**
-	 * Returns raw parameter types.
+	 * Returns {@link MethodParamDescriptor method parameteres}.
 	 */
-	public Class[] getRawParameterTypes() {
-		return rawParameterTypes;
+	public MethodParamDescriptor[] getParameters() {
+		return parameters;
 	}
 
 	/**
-	 * Returns raw parameter component types. Returns <code>null</code>
-	 * if data does not exist.
+	 * Returns number of parameters.
 	 */
-	public Class[] getRawParameterComponentTypes() {
-		return rawParameterComponentTypes;
+	public int getParameterCount() {
+		return parameters.length;
 	}
 
 	// ---------------------------------------------------------------- getter/setter
@@ -169,16 +177,12 @@ public class MethodDescriptor extends Descriptor implements Getter, Setter {
 
 	@Override
 	public Class getSetterRawType() {
-		return getRawParameterTypes()[0];
+		return getParameters()[0].getRawType();
 	}
 
 	@Override
 	public Class getSetterRawComponentType() {
-		Class[] ts = getRawParameterComponentTypes();
-		if (ts == null) {
-			return null;
-		}
-		return ts[0];
+		return getParameters()[0].getRawComponentType();
 	}
 
 	// ---------------------------------------------------------------- toString